refactor: fetch문들 feches 디렉토리 생성하여 별도 보관, 모듈 이름 변경 및 정리
