+++ client/src/components/Breadcrumb.tsx
@@ -1,4 +1,4 @@
-import { usePathArray } from '@utils/usePathArray';
+import { usePathArray } from '@hooks/usePathArray';
 import { ReactElement } from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
