+++ client/src/hooks/usePathArray.ts
