+++ client/src/pages/CategoryPage.tsx
@@ -3,7 +3,7 @@ import { useQuery } from 'react-query';
 import CategoryTable from '@components/Tables/CategoryTable';
 import SideButton from '@components/SideButton';
 import { handleOnError, onSelect } from '@utils/common';
-import { getCategory } from '@utils/useCategory';
+import { getCategory } from '@fetches/category';
 
 function CategoryPage(): ReactElement {
   const { data, isLoading } = useQuery('category', getCategory, {
